requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
werkzeug
python-dotenv
